training/src/make_dataset/requirements.txt
pandas~=2.2.3
numpy~=2.1.3
lakefs~=0.7.1
confluent-kafka~=2.6.1
kafka-python-ng~=2.2.3
scikit-learn~=1.5.2
fsspec~=2024.10.0
lakefs-spec~=0.11.0
pydantic~=2.10.2